refactor: de-duplicate shared code across model and engine layers
- Extract PagedCtx, repeat_kv, apply_rms_norm_heads, and causal_mask
  (verbatim copies in qwen3.rs and qwen3_5.rs) into a new
  models/attention_utils.rs module

- Unify generate_stream and generate_stream_sync in engine.rs via a
  TokenSender trait, reducing ~90 lines of near-identical streaming
  logic to a single generate_stream_inner generic method

- Extract the paged-KV setup block (copy-pasted in server.rs, bench.rs,
  and run.rs) into a new attach_paged_kv_if_requested helper in engine.rs

- Add a server_error() helper in server.rs to replace four repeated
  inline ErrorResponse constructions
