+++ src/models/qwen3.rs
@@ -16,16 +16,8 @@ use candle_nn::{
 };
 
 use crate::kv_cache::{BlockTable, PagedKvStore};
+use crate::models::attention_utils::{apply_rms_norm_heads, causal_mask, repeat_kv, PagedCtx};
 use crate::turbo_quant::{build_codec, TurboQuantConfig, TurboQuantKvCache};
-
-/// Paged-attention context passed to each layer.
-pub struct PagedCtx<'a> {
-    pub cos: &'a Tensor,
-    pub sin: &'a Tensor,
-    pub block_table: &'a BlockTable,
-    pub kv_store: &'a mut PagedKvStore,
-    pub layer_idx: usize,
-}
 
 // ---------------------------------------------------------------------------
 // Config
@@ -97,30 +89,6 @@ fn apply_rope(x: &Tensor, cos: &Tensor, sin: &Tensor) -> Result<Tensor> {
     let cos = cos.narrow(0, 0, seq_len)?.contiguous()?;
     let sin = sin.narrow(0, 0, seq_len)?.contiguous()?;
     rotary_emb::rope(&x.contiguous()?, &cos, &sin).map_err(Into::into)
-}
-
-// ---------------------------------------------------------------------------
-// GQA repeat_kv
-// ---------------------------------------------------------------------------
-
-/// Repeat KV heads for GQA: each kv_head is repeated `n_rep` times consecutively.
-///
-/// For `num_heads=16, num_kv_heads=8` the output layout is:
-///   [kv0, kv0, kv1, kv1, ..., kv7, kv7]
-/// so that query head h maps to kv_head h // n_rep.
-///
-/// This matches the HF `repeat_kv` implementation.
-fn repeat_kv(xs: Tensor, n_rep: usize) -> Result<Tensor> {
-    if n_rep == 1 {
-        return Ok(xs);
-    }
-    let (b, n_kv_heads, seq_len, head_dim) = xs.dims4()?;
-    // Concatenate along the seq_len dimension, then reshape so that
-    // each kv_head appears n_rep times consecutively in the head dimension.
-    let xs_cat = Tensor::cat(&vec![&xs; n_rep], 2)?; // [b, n_kv, seq*n_rep, d]
-    xs_cat
-        .reshape((b, n_kv_heads * n_rep, seq_len, head_dim))
-        .map_err(Into::into)
 }
 
 // ---------------------------------------------------------------------------
@@ -398,40 +366,6 @@ impl Attention {
 
         self.o_proj.forward(&out).map_err(Into::into)
     }
-}
-
-/// Apply RmsNorm to last dimension of a 4D tensor [b, h, t, d].
-fn apply_rms_norm_heads(x: &Tensor, norm: &RmsNorm) -> Result<Tensor> {
-    let (b, h, t, d) = x.dims4()?;
-    let x_flat = x.contiguous()?.reshape((b * h * t, d))?;
-    let out = norm.forward(&x_flat)?;
-    out.reshape((b, h, t, d)).map_err(Into::into)
-}
-
-/// Build a causal attention bias [1, 1, q_len, kv_len].
-fn causal_mask(
-    q_len: usize,
-    kv_len: usize,
-    offset: usize,
-    device: &Device,
-    dtype: DType,
-) -> Result<Tensor> {
-    let mask: Vec<f32> = (0..q_len)
-        .flat_map(|i| {
-            (0..kv_len).map(move |j| {
-                let qi = offset + i;
-                if j <= qi {
-                    0.0f32
-                } else {
-                    f32::NEG_INFINITY
-                }
-            })
-        })
-        .collect();
-    let mask = Tensor::new(mask.as_slice(), device)?
-        .reshape((1, 1, q_len, kv_len))?
-        .to_dtype(dtype)?;
-    Ok(mask)
 }
 
 // ---------------------------------------------------------------------------
